Uploading code package for stitching SSID in Matlab.
Code was delevoped on basis of Lars Buesing's SSIDforPLDS code.
assignOpts.m, ssidSVD.m, vec.m by and large are still 1:1 copies,
post_process.m is extracted from the main script of SSIDforPLDS.

generateCovariancesFP was changed to allow for non-observed entries
in the covariance matrices as observed when stitching data.

soft_impute.m is a simple implementation of an algorithm for
missing data SVD as published in 2010 by Mazumder et al.

iterSSID.m and run_stitchingSSID_experiment.m are implement an
early implementation for extending SSID to the stitching scenario.