Uploading code package for stitching SSID in python. Code was originally developed on basis of Lars Buesing's SSIDforPLDS code. The SSIDforPLDS code was translated to python by Artur Speiser. The code was then further extended to allow applying SSID to the stitching of data sets.
Added the dependency of the 'control' module, which imports Matlab
functionality for solving discrete Riccati equations also for the
non-stabilizing case (which seems to be beyond the scope of scipy).

The translated code seems mostly bug-free, but a little further
testing will be needed to be sure.

There are several shortcomings of the stitching SSID method
(mostly caused by degeneracies of the LDS model representation)
for which practically any solution (even if quite hacky)
would be a big improvement at this point.