Major "bug-fix": the values for stationary covariance Pi are problematic.
Deleted all lines of code where the values for Pi as returned by ssdiSVD
were later on mixed by solutions of the Lyuapunov equation Pi = A*Pi*A'+Q,
as computed by dlyap.

Apparently, the heuristic applied in lack of a proper DARE solver is quite
often _very_ poor, and the Pi returned by ssidSVD indeed does not suffice the
Lyapunonv equation overly well.
Mixing the Pi's as returned by ssidSVD and those from dlyap in many cases
disrupts the iterative scheme.

The DARE solver imported from the control module currently still cannot be
used. It requires the FULL instantaneous covariance matrix to compute a
full-rank R. On the long run, a DARE solver for the general (non-stabilized)
case that assume diagonal R is needed!