Stacked gradients for A,B,C together to learn all at the same time.
Seems to work right out of the box, both for (small) fully observed
systems and with missing parts. So far only check (p,n) = (5,3)
systems with full or 1 overlap (the latter would not be stitchable
just from the instantaneous covariance matrix, i.e. we need the
Hankel structure).

Will need to do further testing.
If results hold, there will also be some algorithmic improvements to
be made; in particular the update of A is gruesome at the moment.