Fixed bug in function to compute derivate w.r.t. A for full Hankel matrices.
Another symmetry issue.
I had missed a transpose in the equation for the gradient for A, which did not
matter for the very first Hankel block, as there the relevant parts of
the gradient only rely on A^0 = I, which is symmetric.

Also completely rewrote the algorithm for computing the gradient w.r.t. to A.
By re-arranging the sums that form the gradient, one can avoid computing
dA^m/dA (a four-axis tensor). It is very basic, could have seen that earlier.