Wrote code for the partially observed case for Hankel L2 coordinate ascent.
The solution comprises a fixed-point iteration to solve for cov(x_{t+m},x_t)
for a given C and observed time-lagged covariance matrices Q(m).
Convergence of this iteration however is fairly fast.

Gradients w.r.t. C for the partially observed case are unchanged relative
to the fully observed case once the fixed-point iteration is done.

Code works on medium-sized problems, as before. For large problems (n > 10
at the moment), the gradient descent does not move a single step, and
instead wrongly declares any initialisation to be a local minimum.
Same as for the full descent w.r.t. A,B,C.

Current best bet is that it is a numerical issue, as for increasing n,
both the size of the involved latent matrices, as well as the minimal
Hankel size, scale with n^2 - there is quadratically more numbers to
compute, and each of them is a sum over quadratically more terms.