Added option to start fixed-point iteration from largest subpopulation.
The fixed-point iteration is only needed in the case of partial
observations. When ignoring all but a single subpopulation, we
are back to the fully observed case and can solve for A^m Pi
analytically. Requires the given population to be large enough to
well estimate the latent covariance matrices only from it.
Then the fixed-point iteration can be run to refine the result
for the whole population.