Implemented alternative solution to 'fixed-point iteration'.
Analytic solution using the observed covariance blocks.
The previous analytic solution (solving system of lin. equations) relied
on the non-observed blocks and required the pesudo-inverse of C.

In general, the fixed-point solution can be faster than the solution of
the system of linear equations if only looking at single time-lags, as
the computation in n^2 - by - n^2 spaces using the kronecker product seem
horribly inefficient.