Further work on identifying why solving for Pi does not work.
Narrowing things down to the first-order solution for A from
A*[X_1, X_2, ... ] = [X_2, X_3, ...]
probably being to poor in quality to use for A^m Pi for large m.

Might have to abandon computing the latent time-lagged covariances
X_m solely from powers of A and from Pi, and instead go with a
'first-order linearisation' of the latent covariance matrices.