Implemented 'intermediate' method for dealing with dynamics matrix A.
We get A from least squares on shifted versions of the extracted time-
lagged covarariances. Instead of using these potentially pretty bad
estimates to get all included time-lags (requiring potentially quite
high powers of A to be used), we just time-shift the latent covarances
once with A and use these as with the non-linear approach.
Guarantees that all latent covariances are at least somewhat related
to the extracted A.

Currently still somewhat fiddly - can explode for n > 5.
Might need further debugging.