Commiting current version before trying out some re-parametrisations.
Might be that the is a bug in the representation of the latent covariances.
Python's reshape() is not the same as the vec() operator, and there
might be an issue with the twisted represenation between the code parts for
extraction of latent covs and the extraction of A from these latent covs.