Started (and commented out) code for 'proper' handling of the temporal observation scheme.
As of now, 'partial observation' is modeled by a mere mask over the covariance matrix.
For 'infinite' data this is fine, but for finite data this is problematic.
What we should actually do is to chunk the time series into intervals of constant
observed populations, compute the covariances for each of these and then combine them.

A currently open issue with the latter is what to put for the desired unobserved
covariance entries. Will need to further think into this.