Continued trimming memory consumption of current implementation.
We wish to go big with this algorithm.

Removed
- copies of observation-masked version of covariances used only for fitting.
  Now using the full un-masked version for fitting and result evaluation
  (the code only looks at 'allowed' entries of the matrices, anyways)
- tracking of superfluous masks (specifically for overlap and cross-overlap)
  that can easily be computed at demand.

Rewrote the computation of the linear system equations for the solution for X_m
from C, D = diag(vec(Om)) and the time-lagged observed covariance matrices.
Previously would compute a kronecker matrix that is p1^2 x n^2, where p1 is
the size of the largest subpopulation.

Might have to do more trimming in the future, but for now my laptop can do
p=10.000, n=20, k+l-1=5 temporal delays .