Fixed extraction of A, Pi along the overall descent (vs. just at the end).
Also introduced a 'final' pass over R at the end of the fitting.
This is necessary as we initialise the R[i] during SGD with size-p mini-batches
only if column/row i is chosen. Thus some R[i] that have not been update in the
last few (hundred) stochastic gradients may be vastly outdated by the termination
of the algorithm, given that C is continuously updated all the time.