Standardised 'chunking' of large matrices into smaller blocks that fit into memory.
Computation of the target loss function and solving for the latent covariances X_m
both requires dealing with the full time-lagged covariances Qs[m], which can only
be handled on little chunk of the whole matrix at a time once p grows sufficiently
large.

Wrote a new function that does this divide&conquer approach and that will be
used for such 'chunking' wherever it pops up in the code. Replaced 'local'
solutions within code that does computation of loss and solving for X_m.